pyproject.toml
[project]
name = "qa-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "allure-pytest>=2.15.0",
    "playwright>=1.55.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-playwright>=0.7.1",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["utils"]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "ruff>=0.14.2",
]

[tool.ruff]
line-length = 120
exclude = ["tests/resources/*"]

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = ["E501", "B008", "B904"]  # B008: function calls in defaults (FastAPI pattern), B904: raise from

[tool.ruff.format]
docstring-code-format = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--alluredir=allure-results",
    "--strict-markers",
    "--disable-warnings"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "ui: marks tests as UI integration tests"
]
